move date functions to fetchUtils
